graph-datasources/graph-sec-F4-filings/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-cloud-storage>=3.7.0
spacy>=3.8.0
spacy-curated-transformers==0.3.1
gqlalchemy>=1.4.0
